Refactor all things. Adds more packages to better organize things.
