+++ src/main/java/com/lol768/battlekits/commands/CommandKitCreation.java
@@ -1,5 +1,6 @@
-package com.lol768.BattleKits;
+package com.lol768.battlekits.commands;
 
+import com.lol768.battlekits.BattleKits;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
