+++ src/main/java/com/lol768/battlekits/listeners/DeathEvent.java
@@ -1,5 +1,6 @@
-package com.lol768.BattleKits;
+package com.lol768.battlekits.listeners;
 
+import com.lol768.battlekits.BattleKits;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
