+++ src/main/java/com/lol768/battlekits/listeners/TagHandler.java
@@ -1,4 +1,5 @@
-package com.lol768.BattleKits;
+package com.lol768.battlekits.listeners;
+import com.lol768.battlekits.BattleKits;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
