+++ src/main/java/com/lol768/battlekits/utilities/BukkitMetrics.java
@@ -25,7 +25,7 @@
  * authors and contributors and should not be interpreted as representing official policies,
  * either expressed or implied, of anybody else.
  */
-package com.lol768.BattleKits;
+package com.lol768.battlekits.utilities;
 
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.YamlConfiguration;
